fix: stats' é
